results

+++ lib/utils/ProfilePic.dart
@@ -20,27 +20,27 @@ class ProfilePicture {
     final StorageUploadTask task = ref.putFile(file);
   }
 
-  Future<Null> downloadFile(String uid, bool cache) async {
-    File file;
-    if (cache == true) {
-      final Directory tempDir = Directory.systemTemp;
-      File('${tempDir.path}/$uid').delete();
-      file = File('${tempDir.path}/$uid');
-    }
+  Future<File> downloadFile(String uid, bool cache, bool hasProfilePic) async {
+    if (hasProfilePic) {
+      File file;
+      if (cache == true) {
+        final Directory tempDir = Directory.systemTemp;
+        File('${tempDir.path}/$uid').delete();
+        file = File('${tempDir.path}/$uid');
+      }
 
-    var ref = await FirebaseStorage.instance
-        .ref()
-        .child(uid)
-        .writeToFile(file)
-        .future
-        .catchError((e) {
-      print(e.toString());
-    });
+      var ref = await FirebaseStorage.instance
+          .ref()
+          .child(uid)
+          .writeToFile(file)
+          .future
+          .catchError((e) {
+        print(e.toString());
+      });
 
-    // var storage = await ref.writeToFile(file).future.catchError((e) {
-    //   print(e.toString());
-    // });
-    if (ref != null) {
+      // var storage = await ref.writeToFile(file).future.catchError((e) {
+      //   print(e.toString());
+      // });
       return file;
     } else {
       return null;
